requirements.txt
fastapi
uvicorn
aiohttp
pydantic
pydantic-settings
langchain
langchain_community
langchain-google-genai
langchain-text-splitters
numpy
faiss-cpu
pymupdf
google-generativeai
python-dotenv
tiktoken
pdfplumber
python-docx
beautifulsoup4
email-validator
unstructured
frontend